Fix multiple issues impacting system stability and performance
Correct a PostgreSQL type casting error in SQL queries, adjust AI task scheduling to prevent queue saturation, and prevent redundant retries for backlog items with missing channels.

+++ server/index.ts
@@ -2746,7 +2746,7 @@ httpServer.listen(
                   AND (caption ILIKE '%battlefield%' OR caption ILIKE '%bf6%')
               )
               UPDATE autopilot_queue q
-              SET scheduled_at = NOW() + ((r.rn - 1) * INTERVAL '8 hours') + INTERVAL '1 hour'
+              SET scheduled_at = NOW() + ((r.rn - 1)::int * INTERVAL '8 hours') + INTERVAL '1 hour'
               FROM ranked r
               WHERE q.id = r.id`
         )
